Replacing use of Fakes and Shims with tests that mock the client
Introducing tests that mock the AOSClient using Moq.
Refactoring the code to improve testability.
Adding interface for the AOSClient.
Adding tests for moving and deleting Workloads.
Changed the LoadUserConfig to support loading the config from the testing environment. A copy of the config file must be put into the test assembly to support testing.