If running on LCS-hosted environment, load storage string from web.config
On LCS-hosted environments the Azure storage connection string is not required in the userConfig. In this case, the connection string is loaded from the web.config file. The file is decrypted first.

I do not have access to an LCS-hosted environment, but on my SingleOneBox environment the correct string can be loaded from webconfig, tables are cleared, and the web.config is encrypted afterwards.

#patch
resolves #97