Refactoring (#89)
* Major refactoring of the CLI files

Introducing a DevToolMenu abstract class that has duplicated logic for the menus, simplifying the individual menu classes.
The CLIFramework can now start with any menu as the root menu.
Many nits fixed to reduce messages in VS, making it easier to spot errors.

* Major refactoring of the ScaleUnitManagement and ScaleUnitManagementTests

These are all trivial changes. Most of them are about using "var" instead of an explicit type to improve readability.
Greatly reduces number of messages in VS.

* Update storage account naming and step factory

Also change variable name in step factory back to objects.

* Update .editorconfig

Using var is no longer preferred everywhere.

* Complying to changes in the editor.config

Changing back to explicit naming in most places.

* More explicit naming

* missing usings

* More explicit namings

* Update signature of DevToolMenu.SelectScaleUnitOptions

SelectScaleUnitOptions now take in a list of scale units to display. This gives more flexibility in case the action cannot be performed on any scale unit (e.g. not on the hub). Also, the function now creates the ScaleUnitContext for the chosen scaleUnit.