Improve error handling from AOS requests (#110)
* Improve error handling from AOS requests

ReliableRun class will catch AOS exceptions and try to suggest changes. Other errors pass through. If a suggestion is found, or it does not make sense to retry, the request is not repeated. Every AOS error is shown to the user.
resolves #105
#patch

* More elaborate TSG and proper error handling

* Changes method name

* Catches Exceptions instead of AOSClientErrors

Timeout errors and similar will be caught by reliableRun